crawler/src/pyproject.toml
[tool.poetry]
name = "bingo-crawler"
version = "0.1.0"
description = ""
authors = ["Lucas <lucaspar@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.6"
bs4 = "=0.0.1"
requests = "=2.22.0"
lxml = "=4.4.1"
boto3 = "=1.9.248"
astroid = "=2.3.1"
awscli = "=1.16.258"
beautifulsoup4 = "=4.8.1"
botocore = "=1.12.248"
certifi = "=2019.9.11"
chardet = "=3.0.4"
colorama = "=0.4.1"
docutils = "=0.15.2"
fake-useragent = "=0.1.11"
idna = "=2.8"
isort = "=4.3.21"
jmespath = "=0.9.4"
lazy-object-proxy = "=1.4.2"
mccabe = "=0.6.1"
pyasn1 = "=0.4.7"
pylint = "=2.4.2"
python-dateutil = "=2.8.0"
python-dotenv = "=0.10.3"
PyYAML = "=5.1.2"
rope = "=0.14.0"
rsa = "=3.4.2"
s3transfer = "=0.2.1"
six = "=1.12.0"
soupsieve = "=1.9.4"
typed-ast = "=1.4.0"
urllib3 = "=1.25.6"
wrapt = "=1.11.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
